[BUG] Opening file with "wt" mode is wrong in cygwin
Problem: Opening file with 'fopen(name,"wt")' asks for the automatic
translation of '\n' to '\r\n' when in cygwin. This is wrong as all shell
script generated by gsl will contain '\r\n' which is not recognised by
the shell as a valid end of line.

Solution: Remove the 't' from the write and append mode string. We can
keep it for reading as it will correctly translate windows line ending
in unix line ending (useful for the case where someone edits a file with
a native windows editor).