[boom] Port code from an unrelease version into current
Problem: When generating the boomake files with the available boom
version, the produced files are different from those commited. This
means the file commited were produced with a later version of boom that
wasn't released.

Solution: Reverse engineer the generated code to deduce what is the code
which generated it. Add this code to the current boom version.